Nest remaining components with single parents
