+++ src/web/components/BotDetails/HealthRow/HealthRow.less
@@ -1,4 +1,4 @@
-@import "../../style/stylesheets/vars.less";
+@import "../../../style/stylesheets/vars.less";
 
 .health-row {
     padding: 8px 0 0 0;
