+++ src/web/components/BotDetails/HealthRow/HealthRow.tsx
@@ -1,9 +1,9 @@
 import { useContext } from "react";
-import { JaiaContext } from "../../context/JaiaContext";
+import { JaiaContext } from "../../../context/JaiaContext";
 
-import { NodeTypes } from "../../types/jaia-system-types";
-import { HealthState } from "../../shared/JAIAProtobuf";
-import { Error, Warning } from "../../types/protobuf-types";
+import { NodeTypes } from "../../../types/jaia-system-types";
+import { HealthState } from "../../../shared/JAIAProtobuf";
+import { Error, Warning } from "../../../types/protobuf-types";
 
 import "./HealthRow.less";
