+++ src/web/components/DataOffloadPanel/DataOffloadQueue/DataOffloadQueue.less
@@ -1,4 +1,4 @@
-@import "../../style/stylesheets/vars.less";
+@import "../../../style/stylesheets/vars.less";
 
 .data-offload-queue {
     display: flex;
