+++ src/web/components/SettingsPanel/LayerSwitcherMenu/LayerSwitcherMenu.less
